requirements.txt
flash_attn==2.5.8
numpy==1.24.4
Pillow==10.3.0
Requests==2.31.0
torch==2.3.0
torchvision==0.19.0 # 0.18.0
transformers==4.40.2